+++ focussight/__init__.py
@@ -1,0 +1,6 @@
+"""FocusSight package."""
+
+from .tracker import run_focus_tracker
+from .summary import summarize_file, summarize_directory
+
+__all__ = ["run_focus_tracker", "summarize_file", "summarize_directory"]